Create RomanToInteger
Put any combination of roman numerals and function pushes an integer. This is done by looping through a string and searching for the letter in the rom dictionary. Also, there is an exception case, if the roman numeral is smaller then the roman numeral that is next, the active roman numeral turns to a negative in the sum